move TimeRatio to clients common
